requirements.txt
# Deep Learning
tensorflow>=2.12.0
keras>=2.12.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=1.7.0

# Data Processing
numpy>=1.24.0
pandas>=1.5.0
scipy>=1.10.0

# Image Processing
opencv-python>=4.8.0
Pillow>=9.5.0
albumentations>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# API
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Serialization
joblib>=1.3.0
pydantic>=2.0.0
PyYAML>=6.0

# Jupyter
notebook>=7.0.0
ipykernel>=6.25.0
